Все приложения разнесены по своим проектам
